pwm: qpnp: use the correct parameter to check the change of PWM period
"pwm->period" is an external parameter that will be updated out
of the pwm-qpnp driver unexpectedly. It's improper to use this
parameter to check whether the expected period is different from
the one being used. "chip->pwm_config.pwm_period" is a pwm-qpnp
internal parameter, which always represents the current period,
so use it to fix the problem.

CRs-Fixed: 641179
Change-Id: I455b828c8c9270ef0f32ab3533daa1bc2c7d183d
Signed-off-by: Xu Kai <kaixu@codeaurora.org>