msm: ion: Update the ion cache maintainance operations
The dma_sync_sg_for_* traverses the scatterlist and performs cache
maintainance using the dma_address of each link. If the scatterlist
is already mapped into an IOMMU, the dma_address is mapped into the
IOMMU, and not the cpu virtual address, which is incorrect for
proper cache maintainance. Instead traverse the scatterlist and
call the cache maintainance on each individual page.

CRs-Fixed: 768053
Change-Id: Ia989fb56236d82184a7f581b8ab4f17c0b93a010
Signed-off-by: Neeti Desai <neetid@codeaurora.org>