driver: mdss: forward port driver from 3.4
This isn't a real port, because the driver
are more better than the real 3.10 and are
specified for hammerhead. The driver are
at least, fixed for 3.10, but for compile it
the kernel need to disable "warning" from build
because are some warning, like this:

drivers/video/msm/mdss/mdss_mdp_overlay.c:1504:4: warning:
passing argument 7 of 'msm_iommu_map_contig_buffer' from incompatible pointer type [enabled by default]

but it isn't a real problem, I search the struct for "msm_iommu_map_contig_buffer"
and it is the same as 3.4, and I think in 3.10 don't get in conflict
because 3.10 have a lot of code, that is abesent in 3.4.
At the moment the kernel compile up to "vmlinux"
and the come out a couple error for reference wrong,
but don't in "mdss". Driver are FIXED at least

Signed-off-by: Fenix46 <scarlataemanuele@gmail.com>