[PATCH 5/5] Makefile: remove wrapper that check compiler warnings
This is the original commit:
https://github.com/Fenix46/Kernel-hammerhead/commit/d214aec9ee0c1ef36d24649d0e52f4fad04fe759#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52

This is necessary, because usb driver are 2 warning when build
and at the moment is better to disable the warning check
because are false positive, in fact usb driver work's fine
storage and adb, I have tested in recovery.
I have add this change like a "patch" for usb so, if anyone
will work on 3.10 for hammerhead, will understand that it is necessary.

Signed-off-by: Fenix46 <scarlataemanuele@gmail.com>