drivers: lm3630: Fix backlight reasume path
The interesting thing's is that the bachlight that I have used
is broke, work's at half, because mdss drivers and display
initialize it, but when get in block screen, kernel don't reasume it.
So I have rework the original lm3630_bl file in 3.10, and now kernel
work's perfectly.

Tested it on Mashmallow 6.0.1 ROM

Signed-off-by: Emanuele Scarlata (Fenix46) <scarlataemanuele@gmail.com>